# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Documentation",
        "description": "Store API Postman and Code definitions documentation.",
        "contact": {
            "name": "Backend Engineer"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/categories": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin List Categories",
                "description": "List all categories for admin",
                "operationId": "0a5ce9ac32f2c8c3dedc5e84892bdf42",
                "parameters": [
                    {
                        "name": "depth",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin Create Category",
                "description": "Create a new category. Send as **multipart/form-data** to support image file uploads.",
                "operationId": "ab5ff56942d867e660ac42c379acccf9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "electronics"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Category image file (max 5MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/categories/{category}": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin Show Category",
                "description": "Get a specific category for admin",
                "operationId": "81e81706f51bf0e13ef3835d9b30e78f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin Update Category",
                "description": "Update an existing category. Send as **multipart/form-data** and include `_method=PATCH` field to support file uploads.",
                "operationId": "20002b071ef997033b5ec6ad693d2216",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Method override required for multipart PATCH",
                                        "type": "string",
                                        "example": "PATCH",
                                        "enum": [
                                            "PATCH"
                                        ]
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Category image file (max 5MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin Delete Category",
                "description": "Delete a category",
                "operationId": "75eccfb8117846a724f4624692b4bf8a",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/categories/reorder": {
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin Reorder Categories",
                "description": "Reorder categories",
                "operationId": "9c5a4efdeb49032c45578c2d283ebe25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categories"
                                ],
                                "properties": {
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "parent_id": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "sort_order": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categories reordered"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundResponse": {
                "title": "NotFound Response",
                "description": "Response for 404 not found error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "Response for 422 validation errors",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The email field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "title": "Category",
                "description": "Product category (detail view)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "slug": {
                        "type": "string",
                        "example": "electronics"
                    },
                    "meta_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "properties": {
                            "thumb": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "card": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "banner": {
                                "description": "Only in category detail endpoint",
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "NotFoundResponse": {
                "description": "Not found error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
...
            "ValidationErrorResponse": {
                "description": "Validation error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorResponse"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}